+++ example/main.nim
@@ -1,6 +1,9 @@
 #[
   Created at: 07/04/2021 11:12:25 Sunday
-  Modified at: 07/04/2021 11:51:53 PM Sunday
+  Modified at: 09/17/2021 01:07:50 AM Friday
+
+        Copyright (C) 2021 Thiago Navarro
+  See file "license" for details about copyright
 ]#
 
 {.experimental: "codeReordering".}